detectors/_Rebuff_virtual_machine_requirements/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
bandit==1.7.6
black==23.12.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.9.0
dunamai==1.19.0
exceptiongroup==1.2.0
Flask==3.0.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.13
langchain-openai==0.0.3
langsmith==0.0.83
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.8.0
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
pbr==6.0.0
pinecone-client==3.0.0
platformdirs==4.1.0
pluggy==1.3.0
pyarrow==15.0.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pytest==7.4.4
pytest-rerunfailures==13.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
-e git+https://github.com/protectai/rebuff.git@4d2fe064abf164e7381556d23e48e210080f8afa#egg=rebuff&subdirectory=python-sdk
regex==2023.12.25
requests==2.31.0
rich==13.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
stevedore==5.1.0
tenacity==8.2.3
tiktoken==0.5.2
tomli==2.0.1
tqdm==4.66.1
types-requests==2.31.0.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
Werkzeug==3.0.2
yarl==1.9.4
zipp==3.18.1
